refactor: clean up common type imports
